ai_service/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
distro==1.9.0
fastapi==0.120.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.6.1
google-generativeai==0.8.3
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyodbc==5.3.0
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.49.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.2